the third week task

+++ app/Http/Controllers/ArticlesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 
 use App\Article;
 
+use App\Favorite;
 use App\Http\Requests\ArticlesRequest;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
@@ -14,6 +15,7 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 class ArticlesController extends Controller
 
 {
+
 
     public function __construct()
 
@@ -43,6 +45,7 @@ class ArticlesController extends Controller
     {
 
         $article = Article::findOrFail($id);
+        $article->check_article($article);
 
         return view('articles.show', compact('article'));
 
@@ -54,7 +57,6 @@ class ArticlesController extends Controller
     {
 
         return view('articles.create');
-
     }
 
 
@@ -63,36 +65,23 @@ class ArticlesController extends Controller
     {
 
         $article = Article::create([
-
             'user_id' => auth()->user()->id,
-
             'title' => request('title'),
-
             'body' => request('body')
-
         ]);
 
-
+        //загрузка и сохванение изображения
         if (request()->hasFile('image')) {
-
             $file = request()->file('image');
-
             $filename = $file->getClientOriginalName();
-
             $file->move('images', $filename);//папка для загрузки изображения
 
-
             if ($filename) {
-
                 $artic = Article::find($article->id);
-
                 $artic->image = '/images/' . $filename;
-
                 $artic->save();
             }
-
         }
-
 
         return redirect()->home();
 
@@ -105,8 +94,19 @@ class ArticlesController extends Controller
 
         $article = Article::findOrFail($id);
 
-        return view('articles.edit', compact('article'));
+        //поиск юзера, который первый лайкнул
+        $favorite = $article->favorites()
+            ->oldest()->where(['favorited_id' => $id])
+            ->first();
 
+        if ($favorite != null) {
+            $favoriteUser = $favorite->user_id;
+            if (auth()->id() == $favoriteUser) {
+                return view('articles.edit', compact('article'));
+            }
+        }
+
+        return back()->with('message', 'You can\'t edit this article');
     }
 
 
@@ -115,13 +115,9 @@ class ArticlesController extends Controller
     {
 
         $article = Article::find($id);
-
         $article->title = request('title');
-
         $article->body = request('body');
-
         $article->save();
-
 
         return redirect()->route('article_id', ['id' => $id]);
 
@@ -134,7 +130,12 @@ class ArticlesController extends Controller
 
         $article = Article::find($id);
 
-        $article->delete();
+        if ($article->user_id == auth()->id()) {
+            $article->delete();
+        }
+        else {
+            back();
+        }
 
         return redirect()->home();
